+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Calculation time estimate 
+
 ## [0.6.2] - 2026-02-20
 
 ### Fixed